Add FAILING test for #4
There is a bug where we attempt to encode '\r\n' in an impossible way in
MIXED mode; see https://github.com/delimitry/aztec_code_generator/issues/7

I'm not convinced that the solution from
https://github.com/dlenski/aztec_code_generator/pull/4 is a very clarifying
or complete one.

For now, just add a test which demonstrates the problem, and is marked as
expectedFailure.